"The Great Ensafening": remove most of `unsafe` code; split BDD functionality into `ROBDDPtr` trait and impl; drop canonicalization from SDDs (#98)
This is a really large PR that's hard to write a good summary for. @SHoltzen and I tag-teamed on converting a big chunk of the codebase to rely on more safe code, which often involved using `RefCell`s to dynamically check borrowing rules when necessary (often being the updating of internal state for the manager & pointer classes). One very large change is the addition of lifetimes to almost all code involving BDDs and SDDs.

Along the way, we made some big changes:

1. create new `ROBDDPtr` and `ROBDDBuilder` traits that abstract the BDD behaviour; the default `BddPtr` and `BddBuilder` now implement the trait
2. remove the trait/mixin-based canonicalization work I did earlier 😓 
    - removes the core `Canonicalizer` trait, moving logic back to the `SddBuilder`
    - removing the mixin for the `BackedRobinhoodTable` to use different hashers
    - removing the `SddApply` trait
    - we'll reintroduce the rest of this soon :)

In addition, we've:

- disabled the tests related to semantic hashing
- updated some doctests
- removed `dump_models`, which only I used
- updated the WASM target