Reduce reliance on runtime `SDD` panics  (#99)
This is an (opinionated) PR that tries to reduce public-facing functions that could `panic!` (more on this below), ~~as well as otherwise remove public-facing functions that are unnecessary.~~ have moved quite a bit of this to #104 to make this easier to review!

## shifting enum access errors leftward

In the SDD codebase, I try to remove an overall pattern that looks like this:

```rust
if s.is_bdd() {
  let var = s.topvar();
  // ...
}
```

`topvar()` is a function that panics if it's not called on a BDD. In this code, we *know* that it's not possible for `topvar()` to panic, since we call `is_bdd()` beforehand. But, this isn't encoded into the type system, and it's not guaranteed that public-facing users of our code will know to do this too! So, when possible, I've converted the pattern to

```rust
match s {
  SddPtr::BDD(bdd) | SddPtr::ComplBDD(bdd) => {
    let var = bdd.label(); 
    // ...
  } 
}
```

which, while slightly more verbose, is type-safe! 

The eventual goal here is to remove *all* invocations to functions like `topvar()`, and remove them from the codebase. There are still a few areas where it's used where it's a bit trickier to resolve :( A nice side effect is that error messages are a little bit better, since runtime panics that are still necessary are now "closer" to their function call.