Draft: filter out false subs
This doesn't work because this will always be a cache hit,
which means that `get_or_insert` never inserts. Thus, I need to either:

1. force an insertion (and dealloc the old node?)
2. create a new builder?
3. something else

Forcing an insertion is probably a good idea since I will eventually
need to do something like that for "picking the best" SDD to store
at each allocation decision.